+++ .gitignore
@@ -10,6 +10,7 @@ app_nanos.json
 app_nanosplus.json
 app_nanox.json
 app.json
+Cargo.lock
 
 # Python
 __pycache__/
